relax float equality comparison
The included test seems to lead to an infinte loop, due to a bug in the
`rand` package, version `0.8.5`. The bug is fixed in `rand` version
`0.9`, but that version is not released yet.

Either way, it can be worked around, if care is taken when to sample a
uniform distribution. This commit samples a uniform distribution only if
the absolute difference between the two numbers is greater than 1e-4
parts of the smaller one (1%%). Otherwise, it assumes that the numbers
are the same and just uses the smaller one.